backend/pyproject.toml
[tool.poetry]
name = "dataforseoapi"
version = "0.1.0"
description = ""
authors = ["bohda <bohdangavryschuk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.0"
Django = "^3.1.7"
djangorestframework = "^3.12.4"
django-rest-swagger = "^2.2.0"
psycopg2 = "^2.8.6"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
